requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.5.1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
boto3==1.35.38
botocore==1.35.38
catalogue==2.0.10
certifi==2024.8.30
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
Django==5.1.1
djangorestframework==3.15.2
dnspython==2.6.1
durationpy==0.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
faiss-cpu==1.11.0
fastapi==0.116.1
filelock==3.18.0
Flask==3.0.3
Flask-Cors==5.0.0
flask-lambda==0.0.4
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kappa==0.6.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langcodes==3.4.1
langsmith==0.4.5
language_data==1.2.0
lxml==5.3.0
mangum==0.19.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.10
mypy_extensions==1.1.0
mysql-connector-python==9.0.0
networkx==3.5
numpy==1.26.4
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.3.0
placebo==0.9.0
playwright==1.47.0
pluggy==1.5.0
preshed==3.0.9
propcache==0.3.2
pydantic==2.9.2
pydantic-settings==2.10.1
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.18.0
pymongo==4.9.1
PyMuPDF==1.26.3
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
s3transfer==0.10.3
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
selenium==4.25.0
sentence-transformers==5.0.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.1
srsly==2.4.8
starlette==0.47.1
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tqdm==4.66.5
transformers==4.53.2
trio==0.26.2
trio-websocket==0.11.1
troposphere==4.8.3
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.20.1
zappa==0.59.0
zstandard==0.23.0
